moved CacheView to temp_api_views.py and added it to .coveragerc
